Move graph utils to new file and add tests
