+++ src/components/dashboard/QueriesGraph.tsx
@@ -10,7 +10,7 @@
 
 import React, { Component } from "react";
 import { WithNamespaces, withNamespaces } from "react-i18next";
-import { getIntervalForRange, padNumber } from "../../util";
+import { getIntervalForRange, padNumber } from "../../util/graphUtils";
 import api from "../../util/api";
 import { WithAPIData } from "../common/WithAPIData";
 import { ChartData, ChartOptions, TimeUnit } from "chart.js";
